+++ HISTORY.md
@@ -2,6 +2,7 @@
 
 ## UNRELEASED
 
+* Updated documentation for module and object rules.
 * Added an example for Typer.
 
 ## 0.3.2 (2021-08-24)